Make a single shared lib handling several modules
Since we are loaded in mesos-slave context,
and so be impacted by the single ruby vm per process,
we are making it explicit,
and factorizing some code (more can be done)!
