+++ README.md
@@ -16,7 +16,7 @@ Build Instructions
     $ cd build
     $ cmake ..
     $ make
-    $ ./test_hook ../hook.rb
+    $ ./test_rmodules ../tests/mesos_modules.rb
 ```
 
 Scripting Documentation
@@ -39,4 +39,20 @@ RubyHook is currently expecting the following API from the attached Ruby script:
 
 Both `taskinfo` and `execinfo` are hashes filled from equivalent C++ classes.
 For instance, you can get the name of a task with `taskinfo["name"]` and access
-the labels kv-pairs through `taskinfo["labels"]` as a string-string hash.  
+the labels kv-pairs through `taskinfo["labels"]` as a string-string hash. 
+
+
+Deployement & Configuration
+------------------------------
+
+This libraries' modules supports the parameter `script_path`, which shall be unique for the library.
+In reality the first one process shall win (no other `script_path` will be taken into account).
+
+This is done on purpose to make it explicit that all modules (hook and isolator) will run in the same 
+ruby context.
+
+Example of json configuration file in examples sub-folder.
+Don't forget to add `com_criteo_mesos_RubyIsolator` to the list of slave's activated isolators
+(--isolation="...,com_criteo_mesos_RubyIsolator") 
+
+
