+++ examples/module.json.example
