+++ src/RubyIsolator.cpp
@@ -1,22 +1,12 @@
 
-#undef NORETURN
-#undef UNREACHABLE
-#include <ruby.h>
-
 #include "RubyIsolator.hpp"
-#include "RubyEngine.hpp"
+#include <ruby.h>
 
 #include <mesos/mesos.hpp>
 #include <mesos/module/isolator.hpp>
 #include <mesos/slave/isolator.hpp>
 
-#include <process/future.hpp>
-#include <process/owned.hpp>
-#include <process/process.hpp>
-
-#include <stout/try.hpp>
-#include <stout/option.hpp>
-
+using criteo::mesos::UniqueRubyEngine;
 using criteo::mesos::RubyIsolator;
 using mesos::ContainerID;
 using mesos::ExecutorInfo;
@@ -24,10 +14,8 @@ using mesos::Parameters;
 using mesos::Parameter;
 using mesos::slave::Isolator;
 using mesos::slave::ContainerConfig;
-
-const std::string RUBY_SCRIPT_NAME = "RubyIsolator";
-const std::string PARAM_SCRIPT_PATH = "script_path";
-
+using process::Future;
+using process::Failure;
 
 extern "C" {
 
@@ -44,35 +32,20 @@ VALUE ruby_isolator_prepare(VALUE obj)
 } // /extern "C"
 
 RubyIsolator::RubyIsolator(const Parameters& parameters)
-  :ruby(RUBY_SCRIPT_NAME)
 {
-  std::string scriptpath;
-  foreach (const Parameter& parameter, parameters.parameter()) {
-    if (parameter.has_key() && parameter.key() == PARAM_SCRIPT_PATH && parameter.has_value()) {
-      scriptpath = parameter.value();
-      break;
-    }
-  }
-
-  if (scriptpath.empty()) {
-    throw std::invalid_argument("missing parameter " + PARAM_SCRIPT_PATH);
-  }
-  if (!ruby.load_script(scriptpath)) {
-    throw std::runtime_error(ruby.handle_exception());
-  }
+  UniqueRubyEngine::getInstance().start(parameters);
 }
 
-RubyIsolator::~RubyIsolator(){
-
-}
-
-process::Future<Option<::mesos::slave::ContainerLaunchInfo>> RubyIsolator::prepare(
+Future<Option<::mesos::slave::ContainerLaunchInfo>> RubyIsolator::prepare(
     const ContainerID& containerId,
     const ContainerConfig& containerConfig)
 {
   LOG(INFO) << "RubyIsolator::prepare( " << stringify(containerId) << ", ..)";
 
+  UniqueRubyEngine& ruby = UniqueRubyEngine::getInstance();
+
   // Now call Ruby
+  std::mutex * mutex = ruby.mutex();
   synchronized(mutex){ if (ruby.is_callback_defined(PrepareCallbackName)) {
     
     // Gather params for Ruby
@@ -95,29 +68,29 @@ process::Future<Option<::mesos::slave::ContainerLaunchInfo>> RubyIsolator::prepa
     // call Ruby in a protect env to avoid exception leakage
     int state = 0;
     VALUE result = rb_protect(::ruby_isolator_prepare, h_params, &state);
-    //if (state != 0) {
-    //  return Error(ruby.handle_exception());
-    //}
+    if (state != 0) {
+      return Failure(ruby.handle_exception());
+    }
   }}
 
   return None();
 }
 
-process::Future<Nothing> RubyIsolator::cleanup(
+Future<Nothing> RubyIsolator::cleanup(
     const ContainerID& containerId)
 {
   LOG(INFO) << "RubyIsolator::cleanup()";
 
+  UniqueRubyEngine& ruby = UniqueRubyEngine::getInstance();
+
+  std::mutex * mutex = ruby.mutex();
   synchronized(mutex){
     if (ruby.is_callback_defined(CleanupCallbackName)) {
       int state = 0;
       VALUE result = rb_protect(::ruby_isolator_cleanup, Qnil, &state);
-      //if (state != 0) {
-      //  return Error(ruby.handle_exception());
-      //
-      //
-      //
-      //}
+      if (state != 0) {
+        return Failure(ruby.handle_exception());
+      }
     }
   }
   return Nothing();
