+++ tests/mesos_modules.rb
@@ -16,3 +16,22 @@ def slaveExecutorEnvironmentDecorator(executorInfo)
   executorInfo['command']['environment'].delete('deleted')
   return executorInfo
 end
+
+def isolator_prepare( params )
+	pid = Process.pid
+	File.open('/tmp/ruby_isolator_stdout.txt','a') do |f| 
+    f.puts "<PREP>"
+		f.puts "MyPID=#{pid}"
+    f.puts "Params (#{params.count}):"
+    f.puts "-----------------"
+    f.puts params.to_s
+    f.puts "</PREP>"
+	end	
+end
+
+def isolator_cleanup( params )
+	pid = Process.pid
+	File.open('/tmp/ruby_isolator_stdout.txt','a') do |f| 
+		f.puts "[CLEAN] MyPID=#{pid}"
+	end	
+end
